Fix character classification macros, like isDIGIT
As fully described in the Changes file, this fixes the character
classification macros in all releases to match the same things they do
in modern perls.  EBCDIC is now supported.

isFOO_A synonyms are added for all isFOO classification macros.  In
modern perls, there are other macros that extend things beyond ASCII,
and the "_A" suffix makes it clear that just ASCII is needed.  Programs
should be able to use these and be understood on earlier perls.  The
macros that deal with beyond ASCII are not ported; generally perls
earlier than when the macros were defined are buggy in this range
anyway.

It would be possible to unify all but one macro between EBCDIC and
ASCII, but the result would be overly complicated for ASCII, and I chose
to keep those separate, but I did unify the ones where the new one is
clearer than the old, like isPUNCT using the literal punctuation
characters instead of ranges of ordinals whose meaning is not obvious.