Add an argument to mktodo.pl
This will be the directory in which to place the output files.

This will enable not having to write over one set of file as we go
along.  Previously everything got written to one directory, then the
files were moved, and the directory filled again.

This will also enable passing in info without having to add even more,
mostly redundant, arguments.