devel/mktodo.pl: Don't discard earliest info
When regenerating the files of what came when, the code uses version n
as a base to see what version n+1 added.  Prior to this commit the
information from version 0 was simply discarded.  That meant you had to
have one more perl to compile than what you wanted info for.  That is
wasteful and unnecessary.

This commit changes so that a file is generated for the earliest perl
being generated, containing all the API symbols known to that perl.
Effectively, it assumes that all of them popped into existence when that
version came out.  Words will be changed in the output to indicate that
these symbols have been available AT LEAST since this version.

For many years, the earliest perl that we had info for was 5.003_07.
That meant that the regeneration was run with an even earlier perl.
With this commit, wherever we stop, we have information from it, and
don't have to throw anything away.