devel/scanprov: Handle blead properly
Previously, the code just removed blead from the list of perls it dealt
with.  But this doesn't work, because the algorithm (heuristic,
actually) relies on going through all the versions, and the symbols
erroneously ascribed to blead will instead erroneously be ascribed to
the next lower version.  Instead, this commit changes so that blead is
simply not written.  (The comments detail why the symbols are
erroneously ascribed to the highest version checked.)