devel/scanprov: Ignore comments in scan
As noted in the comments in scanprov, the results could be thrown off by
something being mentioned in a comment in a scanned hdr.  This commit
removes the comments before scanning.

I'm unsure of what else could throw off the results, but I can imagine
there is something that I (and previous maintainers) haven't thought of.