devel/scanprov: Add in exceptions
Things like RETVAL were being listed as not having any portability
information, whereas they are actually valid for all current releases.
Elements like these can't be tested individually because they are too
tightly bound with XS, which is used to do the testing.  Therefore, they
are in a list of exceptions.  This commit causes scanprov to use this
list and insert the items into the proper version files.