devel/scanprov: Store information about all fcns
This makes it convenient for someone to find out in what version
something got added.  I'm not publicizing this just yet, and
ppport.h --api-check won't return it for non-public ones at the moment,
but the information is now there for those who know where to look, and
perhaps for future public consumption.