devel/mkapidoc.pl: Rewrite entirely in perl
The old version, partially in sh, grabbed every file in the directory
structure, looking for apidoc lines.  Unfortunately that included
things, in .git, for example, and those could be out-of-date entries,
with the result that there could be multiple entries with different
prototypes for the same macro.

This commit changes to read only MANIFEST files, and a few generated
files.