devel/mktodo.pl: Stop using a known bad API element
This was a bug which led to some unpredicatble results due to hash key
randomization.  We determined that a function or macro is not suitable
for a given release, but then forget that fact, and try to use it
anyway.

This single line change fixed a bunch of problems.  As I understand, you
couldn't regenerate on newer perls, and still may have had to try
multiple times for success.  I believe this is because of the hash key
randomization in v5.18.  I sort of understand how that could cause this
kind of issue, but I can't explain it easily, and its gone now anyway.