parts/inc/misc: Backport some isFOO_LC macros
A few of this class of macros did not go back very far.  This makes a
reasonable attempt to get things right, but very early versions may have
some wrong answers, but unlikely.

This was complicated by the fact that isascii() and isblank() may not be
available on a given platform.  So this just uses the plain non-locale
version for those very early Perl versions.

I didn't add tests.  It is hard to portably test locales.  The next
commits will backport functions that call these and do have tests.