Check for PaxHeader in the tarball
Fixes GH #183

Detect and reject tarball produced using PaxHeader

Note: we cannot use the word 'PaxHeader' as it or it
will show up as false positive from the tarball itself.

We are using '[P]axHeader] with a -e to avoid it and spell
the word using a space in the warning message.