parts/inc/version: Handle both p7 and p5
There are CPAN modules that use PERL_VERSION and PERL_SUBVERSION to
decided what to do, while assuming everything is Perl 5.  This commit
basically makes version 7 perls look like 5.200.200 to that defective
code, so that it is larger than any p5 version and less than any p7 one.
The number 200 was chosen for that, and for fitting in a byte, in case
some modules store it in a byte, and for having enough head room so that
we can increase it numerious times and still fit into a byte if
something we haven't foreseen crops up.

There are new definitions in p7 to replace the PERL_VERSION-style ones,
but they are not provided by D:P, so as to discourage their explicit use.
Instead, code that needs to make decisions based on the current perl
version needs to instead use the new PERL_VERSION_xx() macros, where xx
is like 'LE', 'NE', etc.  Doing so will work with any future p8, etc.

But the scheme adopted here should make existing code work, as long as
it doesn't have to distinguish between p7 and p5 or between different p7
versions.  Code that does that will need to change to use the new
comparision macros.

This solution was the idea of Todd Rinaldo, IIRC.  Fitting into a byte
was the idea of Paul Evans.