parts/inc/version: Deprecate old usages on clang
The p5 style versioning definitions should be deprecated.  But this is
hard to do since they are #defines used in pre-processors conditionals
As far as I can determine it requires the C99 _Pragma() functionality.
The behavior of that is not well-defined by the C Standard, and as a
result compilers aren't uniform in their acceptance of it.  In fact gcc
doesn't work, but clang does.  So this adds the deprecation only for
clang C99.