scanprov: Find macros in most headers
Given the previous changes, it's pretty easy to expand that so that any
macro defined in a header can be found and indexed.  A special flag is
used for these as being undocumented.  The reason to index them is that
it makes it easy for someone to find when they first came into
existence, which is a chore that is needed sometimes.  Looking through
the list may lead a core developer to see something whose name indicates
it might be of use to her, even though undocumented.  It brings together
in one place all the available macros.