mktodo.pl: Prevent loop
This came about as a result of an error in a perl macro calling another
with the wrong number of parameters.  The other macro is the one
appearing in the error message, so we aren't turning off the right code
for the next time through the loop.  This commit catches that and
doesn't loop.