devel/regenerate: Avoid removing data in partial runs
When regenerating things, often there will come a point where things
break.  This code tries to preserve the data from before the break when
starting up in the middle.  This is helpful for examining the problems.
Of course, the final product needs a full end-to-end run, but this is
helpful in debugging