Devel-PPPort: use inline function for croak_sv without brace groups
The macro for croak_sv when brace groups are not enabled uses
STMT_START/STMT_END. This works when used as a statement, but would
break if used as part of an expression.

To resolve this, change that macro to use a function. As a static inline
function with a namespaced name, it shouldn't cause problems with
namespace pollution.

(cherry picked from commit 90458a2f3a5400195fd1a5d4b17711eff777b8a3)
Signed-off-by: Nicolas R <atoomic@cpan.org>